Cairo: switch to callconv-agnostic implementation
Building Cairo is a huge PITA, and community VB6 efforts to maintain an up-to-date copy appear to be going nowhere.  To give myself additional flexibility when testing Cairo functions in the future, I want to remove the reliance on stdcall-builds entirely.

(Note that, unfortunately, Cairo is still not used anywhere in production PD code because I keep writing faster pure-VB6 implementations of core Cairo features.  I don't think this is anything special on my end - Cairo is just slow for a lot of traditional tasks.  This may change as we get deeper into vector features, but I haven't profiled extensively enough to guarantee one way or another.)