Gradient LUTs: cache final color directly (instead of pulling it from the LUT)
This allows the user to e.g. put a color node at position 99.99, and a transparent node at position 100.0, then use "no wrap" mode to render a fully transparent border around the gradient.  (Relying on direct values from the LUT are problematic this way, as the final LUT position may contain a slight bit of color data from the next-to-last node if its position is very close to 100.0)