Gradient editor: lots o' new features
- new buttons for importing/exporting gradient files
- gamma correction can now be used when calculating gradient blends.  If a gradient contains colors with wildly different luminance values, this may improve its appearance
- improvements to the way noise gradients calculate density
- misc fixes for a number of minor UI gripes

Next on my to-do list is adding a new panel to the editor - a new "front" panel that simply displays a list of saved gradients, with live previews of each.  This will include a nice selection of default gradients, which should make the tool much more accessible to new users.