Gradient editor: new UI, including a dedicated "gradient collection" panel
The gradient editor has seen a massive overhaul, with many new features and improvements to go along with the new dedicated gradient tool.

The first panel of the dialog is now a dedicated "gradient collection" window.  This panel automatically scans for compatible gradient files in the user's /Data/Gradients folder, and populates a custom-drawn list with previews of each.  The user can sort the list by a variety of criteria (including some neat ones like "hue" to group gradients by average color) and everything has been optimized so that gradients are delay-loaded only when absolutely necessary.  (Similarly, gradient previews are populated on-demand so that even with a very large collection, performance is snappy.)

The primary "make your own gradient" panel has also seen UI improvements, including the addition of a dedicated "save to your collection" button for bypassing common dialogs entirely and saving a created gradient straight to the collection.  Similar improvements have been made to the "noise gradients" dialog.

The many new features required a number of new translations - sorry!  Thank you also to Roy K for an updated German language translation in this commit.