Clean up /Resource and /Support folders
I now realize that these folders contain items that haven't been used in PhotoDemon in 5+ years.  Apologies for not cleaning them out sooner.

The /Support folder in particular could use a rewrite of any number of tools (they're all of the "quick-and-dirty" school of coding) but I'm afraid meaningful improvements will have to wait for another day.