libdeflate: clarify CRC32 initialization behavior
When calculating PNG chunk CRCs, we have to include specific header bits in the checksum calculation.  Rather than jump through hoops to allocate header and data in a contiguous memory stream, it's much easier to just calculate CRCs separately, and feed the results of the previous calculation into the next one.

This change basically makes this possible for all cases, including the extremely rare (1 in 2^32) case where a CRC calculation comes out to the same value as the default CRC initialization value.