PNG exporter: add 8 and 16-bit grayscale and grayscale+alpha support
Note that grayscale+alpha color modes have *not* been added to the auto-detect function yet; that function is used across multiple export formats and it will need to be updated to account for PNG's relatively unique support for grayscale + alpha streams.  (In the past, such images had to be written as RGBA instead, because FreeImage also doesn't support grayscale+alpha.)

Said another way, to test the new grayscale+alpha support, you need to manually specify "grayscale" color mode and "full alpha" alpha mode in the PNG export dialog.

I will update the auto-detect function after migrating indexed color mode exports to the new PNG engine (as the auto-detector needs to be updated with specialized 2- and 4-bit detection as well, and it makes more sense to modify it *after* all color modes are running on the new exporter).