pdPNG: add support for indexed mode images (1/2/4/8-bit)
The code is a wee bit ugly and I'll shortly be refactoring it for cleanliness, but this gives a nice benchmark for "at least it works".

I'm interested in adding an option to automatically test whether encoding an e.g. 4-bit image as 8-bit is preferable, because a quick battery of automated tests shows that libdeflate can produce smaller output on 8-bit data with some frequency (especially if the source data is dithered in any way, because packing dithered 4-bit indices into bytes creates a surprising amount of entropy compared to the bare 8-bit data that only uses values 0-15).  While the savings aren't huge (usually < 10%) the likelihood of small image sizes at such bit-depths means that we could perform test runs basically "instantaneously" and thus gain another advantage in the "automatic PNG optimization" arena.