Palette engine: update to support RGBA modes
PD's optimal palette engine (and fast-palette-matcher) now support RGBA modes, which produce custom "optimal" palettes that consider both color *and* opacity.  This is awesome from a PNG export perspective, as we can now produce pngquant-like images on-the-fly, using our own custom code, if the user asks us to forcibly export an 8-bit image when the source data is 32-bit.

PNGquant's results are still higher-quality on most images (as you'd expect), but our code is faster, much easier to integrate into the PNG export chain, and the resulting PNG files are smaller.  If the image doesn't have large hue variances (like most logos, for example), the differences are especially minimal, so this could be a big win for us.