PNG export: sort palette by opacity prior to export
This potentially shrinks file size a bit, as we can write a truncated transparency table.  (Only non-opaque alpha values need to be added to a PNG's palette transparency table; any values past the end of the table are assumed to be opaque, which means that we can cluster all opaque pixels at the end of the palette and skip writing their alpha values entirely.)

While I was here, I did some work to reuse palette decisions across multiple stages of the PNG exporter, which meaningfully improves export performance of 8-bit paletted images.