Fix delay on first-use of keyboard accelerators
This was an incredibly dumb mistake; the accelerator checks to see if a dialog is active before processing a given hotkey.  Unfortunately, one of the checks was mistakenly loading a heavy dialog in order to check its visibility state.

This has been rectified, which makes first use of a hotkey much much faster.  Sorry I didn't catch this sooner!