New central action manager
To implement batch processing, PhotoDemon has long provided a central "Processor" module that accepts a function name and XML parameter list, which it then translates into actual program functions.

Unfortunately, the Processor module also requires some optional pieces of data, meaning you can't just pass it a bare function name and expect it to work.  For example, it needs to know what kind of Undo/Redo data to create for a given function - i.e. does the whole image stack need to be saved (as it would after Flatten) or just the current layer (adjustments/effects) or some other aspect of the image (metadata changes).  These additional requirements are not always obvious, and they have previously been (unwisely) stored only in the main window's code, as its menus are the only way to invoke most program features.

With the forthcoming creation of a program-wide search bar, this needed to change, as I need a straightforward way to map text strings of program functions (e.g. "Blur") to a corresponding program operation, including all optional data like Undo/Redo requirements.

That's what this commit accomplishes.

The new action manager actually lives in the Menus module, which is weird - I know - but convenient because the menu manager already manages a hierarchical list of program operations.  It also has a bunch of localization code already wired up which allows me to map between localized features names and underlying program operations.  I will probably migrate these features to a dedicated Actions module in the future, but for now, this was the easiest way to get this up and running quickly.

I've also started work on a dedicated search bar UI, some of which is included in this commit.  It is not finished, however, because I needed to implement this action manager first, so I had a place to map searched items!