Search panel: get UI to a usable place
There's a lot of testing still to-do, but the new search bar should be largely functional!  A few quick notes on what I have/haven't implemented so far:

- Search results are fully localized (yay!)
- Search generation is nicely perf-optimized, so results are delivered in real-time regardless of language
- Current text matching strategies are rudimentary, but they work across all languages (which was always a must)
- The search bar plays nicely with run-time theme and language changes
- Currently unusable results are hidden; for example, if no images are loaded, adjustments/effects will *not* appear in search results.  I have mixed feelings about this as it's not entirely intuitive, so I may investigate an alternate presentation
- The search results list doesn't do any neat font changes, like bolding the search term where it appears in each entry a la brower search bars.  Mixed-font rendering in win32 is *really* ugly, but I may try to find a workaround as this would be really nice to have.  (I've also implemented the search bar as an owner-drawn list, just in case I can crack this.)
- I've only tested Win 10 at present.  I don't expect problems on old OSes, but I will test soon "just in case" they hold surprises
- As mentioned in previous commits, only menu entries are currently searched.  Things like tool names (e.g. "paintbrush") are *not* searched, but I should probably find a way to add these, along with a nice way to handle search clicks (like flashing the relevant button or something... idk, this is TBD)

There are probably other items I'm forgetting, but I've been at this for a few weeks now and it's been a nasty project to tackle in 5-minute increments!

(Also, because this control was roughly derived from the font dropdown - which is also an owner-drawn control - I took the opportunity to migrate some changes from the regular drop down implementation to the font drop down, as it looked like some old bugfixes never got merged into it.  Apologies for not noticing this until now.)