Search bar: give priority to results that start with the search query
It's a little more complicated than the title suggests, but with this improvement, a search for e.g. "gra" will return this order of results:

1) Gradient tool
2) Adjustments > Monochrome > Monochrome to Grayscale
3) Adjustments > Histogram

While I was here, I also fixed an issue where the user searches for the currently active tool and attempts to activate it anyway.  It will now flash, as expected, and properly hide the dropdown search results box.