Fixes to language editor; updated Simplified Chinese translation
Thank you to contributor ChenLin for an updated Simplified Chinese translation, and for catching a number of bugs in PD's language editor.

I have tried to rectify various language editor issues as best I can, but PD's (very rudimentary) Google Translate capabilities are currently broken in a number of locales.  A "proper" solution would require translating via Google's translate API, but I have neither the time nor the know-how to implement a wrapper for this in VB6.  As such, the auto-translate feature of the language editor may not be useful for the forseeable future.  Contributions and/or fixes from more knowledgeable coders are more than welcome.