Effects > Artistic > Posterize: overhaul underlying code for perf and quality improvements
Relates to #275.  Thank you to @runner111 for pointing me in this direction.

As posterizing has implications elsewhere in the program (e.g. exporting 16-bit-per-pixel data) it was worthwhile to revisit the code and bring up to modern PhotoDemon standards.  A dozen dithering methods are now supported, each with variable strength, and the underlying posterize code has been cleaned up to improve both performance and quality.