Better method for "show location in Explorer" command
This command is available from the image tabstrip (and should probably be made available elsewhere... though I have no idea where to stick it).

Anyway, shelling explore.exe directly, as we previously did, isn't recommended because it won't select the file if an Explorer window is already open in the same folder.  The correct way to select a file via Explorer is via the SHOpenFolderAndSelectItems API (https://stackoverflow.com/questions/15300999/open-windows-explorer-directory-select-a-specific-file-in-delphi), which PD now uses.