Memory-mapped file interface: improvements when writing to an existing file
Instead of the classic "delete file, create new file in its place" approach, memory-mapped files benefit from reusing existing disk allocations as much as possible.

This also fixes a critical bug when attempting to write to an existing file as if it were a new file via a pdStream object.  Many thanks to Hidayat Suriapermana for catching and reporting!