New Effects > Render > Clouds filter
This has been available in Photoshop for decades.  PhotoDemon had all the required pieces already; it was just a matter of exposing them to the user!

One of the cool things this implementation supports is using gradients to render a cloud's colors.  With a built-in gradient like "flames", the results are beautiful!  Gradients with transparency (e.g. neon) are also very cool.  (Full gradient support was one of the things I was waiting on before implementing this filter - better late than never, as usual!)