Move/size tool: manually handle keyboard tab order
(Relates to #277.)

The move/size tool options panel now uses custom code to handle much of its tab-key ordering.  Controls on this panel are grouped vertically (instead of horizontally, like most PD windows) - a deviation that makes me wonder if perhaps the panel needs to be reorganized to match other patterns.  I'm mulling this over.

In the meantime, tab-key ordering (inc. shift-tab-key) is now much more intuive on this panel.

Thank you to @schneis for catching and reporting.