New File > Export > Animated GIF menu
Relates to #270

This provides a user-facing way to test the new animated GIF export feature.  An export settings dialog is still TODO, but in the meantime, this will produce a usable (but large, unoptimized) animated GIF for testing.

I've also modified the animated GIF exporter to perform "safe" saves (e.g. exporting everything to a temp file, validating said temp file, then only overwriting the destination file - if any - if the save was successful).

Thank you also to Roy K for an up-to-date German localization!