New Layer > Split > Layer(s) into image(s) menus
Relates to #278

New menu options provide a way to split either the current layer - or all layers in the current image - into standalone images.  These images can then be edited independently.

To help PD if the user decides to reassemble the images into a single layered image in the future, the current layer index is added to the new image's default name.

Still to-do is an option to assemble multiple images into a single layered image.