New Layer > Split > Image(s) into Layer(s) menu
This is the opposite of the "Split layers into standalone images" feature implemented in the commit prior.  This option will take any *other* open images in PD, and merge them into the currently active image as layers.

The merge code looks for layers with matching names; if found, it will replace the existing layer's contents with the contents of the same-named image.  This provides an extremely convenient way to split a layer out into a separate image, modify it however you want (including adding new layers atop or beneath it!), then automatically merge it back into the original image in its original place, with any/all changes.

(I'm about to add a prompt to enable/disable this, in case it's not wanted - and note that either way, an undo point is created so you can safely revert to a previous state.)

tl;dr - today's commits make it possible to effortlessly switch between editing a group of images as separate images, or a single image comprised of multiple layers.