Animated GIF Import: add frame times to layer names
Relates to #278

This gives the user a way to easily modify frame times using the existing layer UI.  (Just change the layer name, e.g. "Frame 1 (20ms)" to "Frame 1 (100ms)")

The future animated GIF export dialog should provide a way to override this value, or to set a universal value if desired.