File > Export > Animated GIF: finalize work on encoder
Relates to #278

PhotoDemon is now a full-fledged animated GIF exporter.  Any remaining deficiencies come from FreeImage (e.g. forcing local palettes to always use 256 colors) and are not PD's fault.  I'd happily switch to another GIF encoder if one was available, but alas GDI+ doesn't write multiframe GIF files and libraries like giflib would require a ton of work to be usable from VB6.  So it is what it is.

I've rewritten the core logic of the GIF exporter to do nearly everything in a single pass prior to involving FreeImage at all, which provides a large boost to performance.  This involves a bit more memory (and a lot of specialized caching on our end), but the end result is a pretty quick save process relative to other photo editors.

Individual frames are also now optimized "to perfection".  Changes from frame to frame are calculated on-the-fly, with full support for custom disposal methods and local palettes, which allows PD to losslessly import-then-export complicated animated GIFs like this one from Wikipedia:

https://en.wikipedia.org/wiki/GIF#/media/File:SmallFullColourGIF.gif

No work is required on the user's end.  PD - as always - handles optimizations transparently, and you will always get the best quality possible given the input frames.

Exported animated GIF size is very competitive with other software, with the only losses coming from obnoxious FreeImage limitations.  In the future, I may do what I've done with other file formats and write my own encoder.  This would allow for additional LZW optimizations... but at present, I'm not sure it's worth the effort.

Now that work on the encoder is finished, I'm rethinking my decision to stick this feature in the File > Export menu.  I'm seriously considering reworking it to operate in the normal File > Save menu, with automatic detection for animations in the GIF export dialog.  Official plans TBD.