Use spritesheets to greatly reduce GDI object count in animation window
Relates to #278

For animated files with hundreds of frames, pre-caching frames leads to a ton of GDI handles.  This is easily reduced by using spritesheets to share resources; on an animated GIF like the one shared in #278 (200+ frames), this cuts GDI object count in the animation window by well over 90%, with no impact on performance.

Also included in this commit is an updated German localization c/o Roy K.  Thank you, Roy!