File > Export > Animated GIF: add duplicate frame detection
Relates to #278

This is easy enough to handle, at least if the original source is an animated GIF file.  (If we're converting full-color frames to an animated GIF, duplicate frame detection may not work because it's performed prior to quantization - the reasons for this are complicated.)

There are a few extra optimizations I could write, particularly when downsampling full-color animations to GIF format, but as I see these use-cases being extremely rare in real-world usage, I think my time is better spent elsewhere.

Next up, fixing animation previews so that they still work even when layers are not consistent sizes.