Animated PNG importer: tons of bug-fixes for edge-case images
Many thanks to the helpful APNG test suite at this link:

https://philip.html5.org/tests/apng/tests.html

It provides a wide variety of unconventional animated PNGs for testing, and PD now correctly handles every image in the "Valid Image" section.

Note that one thing I have *not* implemented is color management for any frame beyond the first one.  Could I implement this?  Yes.  Is it worthwhile?  idk.  I have yet to encounter an "in-the-wild" animated PNG that defines gamma or chromaticity or full-blown ICC profiles, so I've postponed this for now (as it would add quite a bit of time to the animation import process).