Update README.md to show changes since last stable release
I'm always getting asked about this, and I can't blame people for not wanting to trawl the commit log.  May as well condense the list of big changes in one place.

I feel like I left out a lot of details on the list, but wow - there's still a lot.  Time to start planning the timeline for the next stable release, I guess...