New tool: Clone Stamp
There are still some minor UI issues to address (e.g. proper cursor icons when in "set source point" vs "paint" modes), and at least one brush property is still to-do (sample merged) but the tool itself is otherwise 100% functional and working well!

Clone stamp is more complicated than a regular paint brush because we have to produce a novel brush image "on the fly" using data from the user's selected source image+layer.  (Importantly, PD allows you to use *any* open image and/or layer as the source - not just the one you're painting on!)  Because PD's brush supports a true "hardness" setting, we also have to dynamically mask the brush to emulate the "soft" edges as we go.  This makes the brush much more computationally intensive than a standard single-color brush.

Despite that, VB is more than up to the task, and I haven't even done serious profiling yet.  For a first-draft, this prototype is extremely encouraging - and it makes me excited to target a healing brush in the near future.

Next few days will involve improving UI issues and adding a "sample merged" feature.