Be more aggressive about releasing resources when switching tools
The new clone tool maintains a lot of background temporary resources (to help keep the painting experience fluid), so this was a good opportunity to improveme to the way PD handles temporary resources when switching between on-canvas tools.

It's always a balancing act to maintain enough temporary data to make the app fluid, while not going overboard and keeping *everything* in memory, so I've tried to strike a reasonable balance.  Please let me know if you notice any undesirable behavior (like "stutters") when quickly swapping tools.