Fix viewport bug on first use of pencil tool...
...if the softbrush tool was used immediately before it.

This is a weird bug involving the LOD cache for the viewport when paint tools are active, and I'm still not entirely sure why it only arises when switching immediately from the softbrush to pencil tool.  (Like, why doesn't the clone tool also exhibit this behavior?  No idea.)

Regardless, an easy failsafe is to forcibly clear the relevant LOD cache *after* applying brush results.  This should probably be done anyway to prevent unpredictable tool conflicts in the future, which is why I haven't gone to the trouble of tracking down the exact explanation for the (now solved) bug.

Also included in this commit is an updated German localization c/o Roy K.  Thank you, Roy!