libdeflate wrapper: add compression level validation in GetWorstCaseSize()
libdeflate requires calls to GetWorstCaseSize() to supply a desired compression level.

Unlike zlib, libdeflate does not support a "0" compression level - it always performs at least *some* attempts at compression, and minimum level is 1.

My wrapper was checking for (cmpLevel < 1) when compressing, but *not* when checking worst-case size.  This was causing compression buffers to be allocated to size 0, which obviously caused subsequent compression attempts to fail.