Gaussian Blur - fast approximation: rework radii calculations
Relates to #279

I haven't forgotten about this!  It's just taken me some time to study the math involved and figure out a proper plan of attack.

In this commit, I've reworked the way PD's fastest gaussian approximation filter works.  Thank you to Ivan Kutskir and this link for advice on how to handle this:

http://blog.ivank.net/fastest-gaussian-blur.html

After trying a bunch of different approaches, his is the one I like the most - because the math is approachable (yay!), and also because the similarity to Photoshop is really, *really* good.

As a byproduct of this, PD's "fast" and "precise" modes now produce much more similar results under most radii.

I am now going to attempt to rework edge handling for this same algorithm, so that both it and the IIR filter use the same technique.