Gaussian IIR estimator: quality improvements
This brings the final result of this tool closer to our fast gaussian approximator (which is, in turn, nearly identical to Photoshop's).

That said, I've just discovered that the author of the original function has published an updated version with some quality and performance tweaks, which I'm now going to attempt to integrate into PD.  Those changes may negate these ones, so if you're reading this in the future, please study the commit immediately following this one to make sure these changes are still worthwhile!