Gaussian blur: new (extremely high quality) Deriche IIR approximator
Relates to #279

Well, this is one of the ugliest C translations I've done in awhile.  Many thanks to Pascal Getreuer for his excellent paper, "A Survey of Gaussian Convolution Algorithms":

http://www.ipol.im/pub/art/2013/87/

In that paper, Pascal confirms what was suspected in #279 - that the Alvarez-Mazorra algorithm used by PD's IIR gaussian implementation can deviate strongly from a true gaussian (10+%, even at 3rd-order!).  This is unacceptable, alas.

Fortunately, the same paper suggests an excellent replacement: another time-invariant IIR approach based on Deriche's 1993 paper, "Recursively implementing the Gaussian and its derivatives":

https://hal.inria.fr/file/index/docid/74778/filename/RR-1893.pdf

Pascal adds a number of important updates to Deriche's original method, including significant improvements to edge-handling.  Critically, he also went to the trouble of calculating deviations from a true gaussian for all surveyed approximators, and Deriche's was confirmed to be the most accurate of the bunch, with a deviation of well under 1% at third-order, and less than 0.1% for 4th-order.  (PD's implementation provides 2nd, 3rd, and 4th order implementations, with a current default to 3rd order.)

Like the old IIR method, Deriche's performance is radius-invariant, with no meaningful difference between a radius of 1 or a radius of 500.

I have now modified PD's Effects > Blur > Gaussian Blur tool to use the Deriche algorithm when the "precise" setting is selected.  I may update this tool's UI to directly expose all estimation algorithms, as they all have trade-offs, and it may be worthwhile to allow users to increase the order for AM and Deriche's approximators in particular (for when extreme quality is required).

The new Deriche algorithm is slightly slower than PD's AM algorithm, and about 3x slower than PD's "fast" iterative box blur algorithm.  However, it's still several orders of magnitude faster than a "true" gaussian, with imperceptible differences in standard sRGB space.

Many thanks to @softhorizons for their original detective work on this topic, and for their patience while I worked on a solution.