Fill tool: add preliminary support for macro recording and playback
Relates to #286

This is a quick-and-dirty attempt at allowing fill tool operations to be recorded and played back as part of macros.  On-canvas paint tools were never meant to be compatible with macros (as they introduce a huge amount of complexity), but the fill tool is a bit of an exception as we only need to deal with independent mouse clicks - not more complicated behaviors like mouse strokes.

I'm not convinced this first approach will be bug-free, but it's worth a try!