Dropdowns: manually trigger GotFocus on hovered MouseWheel
This fixes a problem where controls in the main window (e.g. layer blend mode) are changed on Win 10 by hover + mousewheel do not add their changes to the current Undo/Redo chain.  Win 10 allows interaction via MouseWheel when a control does *not* have focus, which is very helpful as it lets the user bypass the need for an initial click.  However, this messes with PD's Undo/Redo tracking as it requires GotFocus/LostFocus pairs in order to know when non-destructive changes occur on controls available in the main window.