New Layer > Transparency > From Luminance tool
(This is an esoteric tool, but I've added it as I was using GIMP to do something similar for a client project.)

This tool provides an easy way to texturize a base layer.  Add the texture you want as a new layer, convert it to black and white (Image > Black and White), then use this new tool with the "gray is transparent" setting.

If you don't have textures on hand, you can try this for a bit of fun to see what the tool does:

- Load an image
- Layer > Add > Duplicate of current layer
- Effects > Edge > Emboss (use default settings, but perhaps bump up depth)
- Adjustments > Black and White
- Layer > Transparency > From Luma (gray is transparent)