Effects > Transform > Perspective: large performance improvements, improved UI theming
This commit pulls in the improvements from 35c583a111bc21ceffc768cb3194fcf814702ecd, cutting the time required for rendering roughly in half.  This greatly improves responsiveness in the UI as well.

I've also reworked the interactive "transform" canvas to use proper UI theme colors.  The end result is a much-improved tool window.

Ideally, the transformed image itself would render in the transform window (instead of just relying on an outline), but that's complicated to do using the current approach... so I'm gonna save that project for another day.