Effects > Artistic > Kaleidoscope: large performance improvements
As with other commits from the past two days, this effect has now been migrated to pdFilterSupport's new, much-faster interface.  It's now twice as fast (roughly) as the old version.

I chose to migrate this effect next as it was the last function still using the old pdFilterSupport.SetPixels() function.  Now that the function is orphaned, I've removed it for good!