Use invisible internal trackers when flagging problematic metadata...
...at export-time.

It's a little complicated to explain this, but by default, ExifTool will try to copy over *all* metadata when saving an image to a new filename.  It's usually very good about knowing which tags should not be written to the file (due to format constraints, for example when saving from JPEG to PNG), but sometimes we need to manually intervene.

Before, I was simply reusing the existing "tag-marked-for-removal" flag that the interactive Edit Metadata window uses.  However, this had the bad side-effect that when a user saves a file to a new format, *then* goes into the metadata editor, they'll see a bunch of tags flagged for removal!  This is despite the fact that those flags may not be required if they save to a different file type later on.

Instead, PD now uses transient flags that are rebuilt on each Save/Save As request.  These flags are internal-only and are not displayed to the user anywhere.  They are OR'd with any user-flagged removal tags when constructing the list of tags to export to the newly saved image.