Start prepping for an 8.0 release
It's been a tough few weeks (for everyone), and it looks like things could stay that way for a while.  Seems like as good a time as any to start working on a new stable release.

PD's nightly builds are in a good place, and I don't want to tackle any large feature additions in the near future.  Instead, I'm focusing on fixing papercuts, and it would be great to get as many people as possible testing the current build(s).

I've got a few more items to work on before upgrading to "beta" status, at which point I'll update photodemon.org with requests for testing.

Many thanks to Roy K for an updated German localization in this commit.