Layer > Crop > menu: change "Pad to image size" to "Fit to canvas"
Relates to #290

I primarily use this option for null-padding a small layer to the full size of the image, but it will also crop the active layer to canvas boundaries.  Renaming it clarifies that it doesn't just add space to the layer canvas - it can remove it, too.

Also, this wording more closely matches the canvas-resize tools in the Image menu, which do the same thing but in reverse (e.g. fit the canvas around a given layer's size).

Thank you to @vatterspun for helping me improve this