Adjustments > Lighting > Gamma: theme UI, large perf improvements
This UI was extremely sluggish before, and I'm not entirely sure why - it seemed to be a GDI+ issue, but somehow pd2D doesn't suffer from the same issue?  Very strange.

At any rate, the new UI is buttery smooth and properly themed on a dark theme (and prettier now, too!)