New Layer > Alpha > Threshold tool
Relates to #300
Relates to #249

This new tool is straightforward: it takes an image's alpha channel, and converts areas of partial transparency to a pattern of fully opaque and fully transparent "dots".  The tool provides full control over dithering with the same options as other palettize tools, but here, they *only* apply to the alpha channel.

This feature is extremely helpful for ICO and GIF exporting, as it provides a way to mimic partial transparency (e.g drop shadows) using a checkerboard of purely on/off pixels.  This feature was also on my to-do list for implementing #249 properly, as it provides a nice way to support layered images and/or text in a "retro" context.

Also included in this commit is an updated German localization c/o Roy K.  Thank you, Roy!