When pasting a new layer, use mouse position for layer x/y offset
Relates to #287
Relates to #303

Thank you to @jpbro for this excellent usability suggestion.

When pasting a new layer using the Ctrl+Shift+V keyboard shortcut, if the mouse is positioned over the image canvas, the new layer will be pasted to the mouse's current coordinates.  This is nice and intuitive, and it works with previous #287 changes (like auto-activating the move tool after a paste) for a much-quicker copy+paste workflow.

If the mouse is *not* currently over the canvas area, or if the Paste event is invoked from the Edit menu, the old logic will still be used (layer gets pasted to the top-left of the current canvas viewport, ensuring it is visible on-screen regardless of canvas zoom or offset).

Also, while here I fixed a bug where certain Paste as New Layer events were creating two Undo nodes instead of one  :astonished: