Splash screen: switch to (faster) layered window approach
Finally knocked this stupid project out!  It's been on my todo list for (literal) years, and it needed to be done before 8.0 releases.

In ancient times (Windows XP), layered windows with alpha effects could carry a large performance penalty.  To avoid this, PD would take a "snapshot" of the current screen before loading, paint the relevant screen region onto the background of its splash screen, then render the program logo atop that opaque snapshot.  This worked well, and was only wonky if the splash screen happened to appear over something animated like a video (because the region behind PD's logo would appear to momentarily freeze).

Anyway, fast forward 20 years and locking the screen for a screenshot is now a very expensive operation due to OS compositor changes.  As such, it's stupid for PD to use this workaround instead of just making a semi-transparent window and painting directly onto it.

I've now rewritten the splash screen to use a proper layered window approach.  This improves cold startup time on Windows 8/8.1 and 10, is a wash on Vista/7, and may incur a slight startup penalty on XP.  I think this trade-off is okay, especially for a program that typically starts in 1-2 seconds anyway... ;)