PNG export: new "use original file settings" option
Relates to #255

It has taken me a long time to add this, sorry!  Also, this feature is only available for PNG files at present because that's a file format that I control end-to-end (e.g. no 3rd-party libraries involved).  That's the only way I can make this work, as libraries like FreeImage can modify palettes internally (like always padding out to 256-colors) without any way to override 'em.

Anyway, the idea here is simple: if a user loads a PNG file, then modifies it, then wants to save it out to file with identical color-depth settings (including palette order), there is now a "use original file settings" toggle that does just that.

Per the request I originally received, this is useful when editing 8-bpp images with fixed palettes.  Sometimes these files have palettes where palette order matters, and reordering the palette - even if it reduces file size - is undesirable.

This was a complicated feature to enable because has to work across the huge matrix of possible PNG settings - some of which we now need to cache when a file is first-loaded, just in case we need to access them again later.  I think I've covered all possible branches, including esoteric ones like HDR images, but please let me know if you encounter any surprises.

In the future, I may be able to extend this to other file formats, but it'll always require custom encoders and decoders to make it work.