Animated screen capture: first test of new frame caching strategy
It works!  Instead of passing frames directly to the APNG encoder, we instead compress them on-the-fly and cache them in-memory.  For typical backgrounds this uses little memory; lz4 can compress a typical application window down to 10% or less of its uncompressed size.

When the recording finishes, we then build the full APNG file.  This actually works better in the end, as I'll be able to do full frame optimization (instead of limiting the engine to fast methods only, on account of needing to run in real-time).

Next up, building a proper capture window.