Animated screen capture: get the "form cut-out" capture rect working
Jeez, this was a mess to make work.  My goal was simple enough - like LICEcap, I want the capture window to have a square "hole" cut out of it, which you can position over the area you want captured.  Simple, right?

No.  Never.  Not when Windows is involved.

I tried three different methods and ended up circling around all of them multiple times, as they all have trade-offs.  Making them work with VB is particularly gnarly as VB likes to take over certain windows messages and do its own thing, which requires constant intervention to make all sides happy.

I finally settled on using a transparent key color and layered windows as PD's solution of choiec.  (I'd say this is the "goldilocks" option in terms of difficulty; SetWindowRgn is what LiceCAP uses and it's easier in many ways, but it forces the window to be drawn un-themed, with a gray border filling in the system drop-shadow area.   On the other end of the spectrum, I also experimented with a full-blown 32-bpp manually rendered window, but that requires manually rendering all non-client window areas, which makes the dialog inadvertently stand out from everything else in PD.... which is just overkill.  I'd also need to make it work on every version of Windows back to XP, so nope, I had to drop it.  My preferred solution is the least of several evils, I think.)

An awful lot of subclassing and API work is required to make this behave correctly on Win 10, and I haven't actually tested other OS versions yet so I can't promise that everything works there.  (It *should* work, in theory, but there are a lot of variables at play.)

While here, I added duplicate-frame-detection to the recorder.  This is (obviously) an incredibly common event during screen captures, and it cuts down on resource usage to simply ignore duplicates when we find 'em.  Frame times are automatically adjusted for these as they occur, regardless of capture frame rate.